+++ {{cookiecutter.project_slug}}/tests/integration/test_docker_container.py
@@ -18,27 +18,7 @@ import pytest
 _FOLDER_NAMES = ["input", "output", "log"]
 _CONTAINER_FOLDER = Path("/home/scu/data")
 
-@pytest.fixture
-def docker_client() -> docker.DockerClient:
-    return docker.from_env()
 
-@pytest.fixture
-def docker_image_key(docker_client: docker.DockerClient) -> str:
-    image_key = "simcore/services/{%- if cookiecutter.project_type == 'computational' -%}comp{%- elif cookiecutter.project_type == 'dynamic' -%}dynamic{%- endif -%}/{{ cookiecutter.project_name.lower().replace(' ', '-') }}:latest"
-    docker_images = [image for image in docker_client.images.list() if any(image_key in tag for tag in image.tags)]
-    return docker_images[0].tags[0]
-
-def _is_gitlab_executor() -> bool:
-    return "GITLAB_CI" in os.environ
-
-@pytest.fixture
-def temporary_path(tmp_path: Path) -> Path:
-    if _is_gitlab_executor():
-        # /builds is a path that is shared between the docker in docker container and the job builder container
-        shared_path = Path("/builds/{}/tmp".format(os.environ["CI_PROJECT_PATH"]))
-        shared_path.mkdir(parents=True, exist_ok=True)
-        return shared_path
-    return tmp_path
 
 @pytest.fixture
 def host_folders(temporary_path: Path) -> Dict:
